+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
     instead the (equivalent) program with `RESET` omitted (@kilimanjaro, gh-1118).
 -   Broadened the scope of `flake8` compliance to the include the `examples` and
     `docs` directories, and thus the whole repository (@tommy-moffat, gh-1113).
+-   `DEFGATE ... AS PAULI-SUM` is now supported (@ecpeterson, gh-1125).
 
 ### Bugfixes